requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
prompt-toolkit==1.0.14
PyAudio==0.2.11
Pygments==2.14.0
PyInquirer==1.0.3
python-dotenv==1.0.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
scipy==1.10.1
six==1.16.0
SpeechRecognition==3.8.1
spotify-cli==0.3.11
spotipy==2.22.1
tabulate==0.9.0
tqdm==4.64.1
urllib3==1.26.14
wcwidth==0.2.6
yarl==1.8.2
